+++ example/lib/main.dart
@@ -34,14 +34,16 @@ class FlutterReduxApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
-      theme: new ThemeData.dark(),
-      title: title,
-      home: new StoreProvider<int>(
-        // Pass the store to the StoreProvider. Any ancestor `StoreConnector`
-        // Widgets will find and use this value as the `Store`.
-        store: store,
-        child: new Scaffold(
+    // The StoreProvider should wrap your MaterialApp or WidgetsApp. This will
+    // ensure all routes have access to the store.
+    return new StoreProvider<int>(
+      // Pass the store to the StoreProvider. Any ancestor `StoreConnector`
+      // Widgets will find and use this value as the `Store`.
+      store: store,
+      child: new MaterialApp(
+        theme: new ThemeData.dark(),
+        title: title,
+        home: new Scaffold(
           appBar: new AppBar(
             title: new Text(title),
           ),